📦 feat: created context provider component
